#2231 move dummy vp to own class
